refactor(desktop): split ChatInput inner components into own files
extract ProviderUsagePill, PreflightPill, and ModelPicker (with its
private PickerSection/PickerDivider/PickerRow helpers) from
ChatInput.tsx into dedicated files. ChatInput drops 739 -> 371 lines.
no behavior change.

Closes #403
